Identidad nueva: negro, estrellas y un solo acento
El sitio se veia barato y la causa era el color, no las piezas. El cian neon
a maxima saturacion sobre negro vibra en el borde de las letras y hace que
todo -bordes, cifras, iconos, botones- pida atencion a la vez.

Ahora casi todo es escala de grises y el color queda para lo poco que
importa: azul frio como acento, violeta solo como segundo tono de un
degradado, y el oro reservado a lo que se gana.

  - Paleta #07080a / #5b9dff / #a78bfa. Se repintaron los 18 ficheros que
    tenian el cian escrito a mano: media web en un color y media en otro es
    peor que no haber tocado nada.
  - Inter + Sora en lugar de Barlow + Oswald. Oswald es una condensada de
    marcador de estadio, y condensada + neon + mayusculas es la formula del
    cartel de gimnasio.
  - El fondo pasa a ser UNA cosa: un campo de estrellas en tres capas de
    profundidad. Antes habia chispas, tres halos y una rejilla moviendose a
    la vez, compitiendo con el contenido que tenian encima.
  - ELITE COIN en el menu deja de ser un rectangulo en degradado detras del
    enlace (se leia como un borde pegado) y pasa a ser una pastilla dorada.

El sitio no compilaba
---------------------
mi/page.tsx tenia un <img> sin cerrar, un useState FUERA del componente -que
revienta el prerender entero- y dos imports ausentes; comunidad/page.tsx
llamaba a getProgreso sin importarlo. Por eso online seguia el diseno viejo.
RANK_COLORS vivia escondido dentro de la pagina de torneos y ahora es un
modulo propio: un dato que usan dos pantallas no puede ser privado de una.

Tokio IA: deja de responder a cualquier cosa
--------------------------------------------
El catalogo agrupaba modelos ("iPhone 11/12/13" era UNA ficha), asi que dos
telefonos distintos daban los mismos numeros por diseno. Emparejaba con la
primera regex que casara, de modo que "iphone 13 por mxa" caia en la regla
del 13 normal sin avisar. Y lo que no reconocia lo respondia igual.

Ahora son 63 fichas, una por modelo, con nombre canonico y resolucion difusa
(tildes, "doce" -> 12, "por mxa"/"promax" -> "pro max"). Con nota baja NO
inventa: sugiere. El eje que separa a dos telefonos de pantalla identica son
los fps sostenidos, que es un dato real; inventarles diferencias de pantalla
habria sido mentir.

Marcador de la portada
----------------------
Top Elite Coin fijo el primero y los demas barajados en cada visita. De 10
paneles a 19, usando lo que el censo antes no leia: % de headshot real, % de
top 10, dano por partida, revividas, nivel y puntos de cada modo.

+++ supabase/mercado_calibracion.sql
@@ -179,6 +179,11 @@ GRANT EXECUTE ON FUNCTION market_push(text, text, bigint, numeric) TO anon, auth
 GRANT EXECUTE ON FUNCTION house_trade(text, bigint) TO authenticated;
 
 -- Como queda el empujon de cada cobro.
+-- El casteo a `numeric` no es un adorno: `ln()` sobre un entero devuelve
+-- `double precision`, y `round(double precision, integer)` NO EXISTE en
+-- PostgreSQL -solo la version de dos argumentos, sin decimales-. Sin el casteo
+-- el script entero falla en esta ultima linea y no se aplica NADA.
 SELECT c AS coins,
-       round(LEAST(0.005, 0.00015 * ln(c + 1)) * 100, 3) || ' %' AS mueve
+       round(LEAST(0.005::numeric, 0.00015::numeric * ln((c + 1)::numeric)) * 100, 3)
+         || ' %' AS mueve
   FROM (VALUES (10),(60),(200),(500),(1500),(4000),(6000)) v(c);
